// src/app/shared/models/usuario.ts
export class UsuarioModel {
    email: string;
    password?: string;
    nombre: string;
    estado: string; 
    fechaCreacion?: Date;
    uid?: string;
    rol: string; 
    fechaModificacion?: Date;
    usuarioCreacion?: string;
    usuarioModificacion?: string;
    displayName?: string;
    photoURL?: string;
    phoneNumber?: string;
    id?: string;
    fechaPago?: Date;
    fechaVencimientoPago?: Date;
    nombreImagen?: string;
    tipoPlan?: string; 
    codigoS?: string;
    codigoR?: string;
    terminoC?: boolean;
    tipoPago?: string;
}